feat: context menu on right-click
